chore: created components folder and moved sectionCard to it
