--- app/components/Views/TrendingView/SectionHeader/index.ts
@@ -1,1 +1,0 @@
-export { default } from './SectionHeader';
